\docType{data}
\name{configs}
\alias{icils_conf}
\alias{piaac_conf}
\alias{pirls_conf}
\alias{pirls2016_conf}
\alias{pisa_conf}
\alias{timss4_conf}
\alias{timss8_conf}
\alias{timss2015_conf}
\title{Config files for intsvy studies}
\format{A list with three components - input, variables and parameters.}
\usage{
pisa_conf
}
\description{
Each config file describes detailed study meta-data. Such meta data defined names of columns with weights, type of weighting, number of plausible values and other study parameters. Most of intsvy functions require such config objects.
}
\keyword{datasets}
