\name{country.ug}
\docType{data}
\alias{country.ug}
\title{Country abbreviations and codes}
\description{
The dataset contains the names, abbreviations, and ISO codes for the countries participating in PIRLS and TIMSS. These labels are not included in the original datasets are necessary for selecting countries and printing outputs.}
\usage{country.ug}
\format{A data frame}
\source{The PIRLS/TIMSS User Guides}
\keyword{datasets}
