\name{iea.country}
\alias{iea.country}
\docType{data}
\title{IEA country abbreviations and codes}
\description{Contains IEA participating countries, names, abbreviations, and code 
}
\usage{data(iea.country)}
\format{
  A data frame with 108 observations on the following 3 variables.
  \describe{
    \item{\code{Country}}{a character vector}
    \item{\code{ISO}}{a character vector}
    \item{\code{Code}}{a numeric vector}
  }
}

\source{PIRLS and TIMSS User Guides
}
\examples{
data(iea.country)
str(iea.country) 
}
\keyword{datasets}
