\name{iea.data.id}
\alias{iea.data.id}
\docType{data}
\title{Meaning of IEA data file names}
\description{A table with abbreviations in file names and corresponding instruments
}
\usage{data(iea.data.id)}
\format{
  A data frame with 8 observations on the following 3 variables.
  \describe{
    \item{\code{Test}}{a character vector}
    \item{\code{Instrument}}{a character vector}
    \item{\code{Abv}}{a character vector}
  }
}

\source{
TIMSS and PIRLS User Guides
}

\examples{
data(iea.data.id)
str(iea.data.id)
}
\keyword{datasets}
