\name{intsvy-package}
\alias{intsvy-package}
\alias{intsvy}
\docType{package}
\title{Analysis of international large-scale assessment datasets (PIRLS, TIMSS, PISA)}
\description{The package provide useRs with tools for reading and analyzing datasets from three international assessments (i.e., PIRLS 2006, TIMSS 2007, and PISA 2009). To read the data, it includes functions that view labels, allow the useR to select variables and countries, merge student and school files, and finally import the data for analysis into R. For the analysis, it includes functions for producing frequency tables and calculating correlations, means, and regressions that handling plausible values and the complex sample design (i.e., replicate weights)}
\details{
\tabular{ll}{
Package: \tab intsvy\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-11-23\cr
License: \tab GPL-2\cr
}
}
\author{Daniel Caro <daniel.caro@education.ox.ac.uk>}
\references{Technical reports of the corresponding surveys}
