\name{is.vlabeled}
\alias{is.vlabeled}
\title{
Utility function
}
\description{
Looks for value labels in data
}
\usage{
is.vlabeled(x)
}
\arguments{
  \item{x}{
An element from a list
}
}
\details{
A utility function for PISA data importation
}
\value{
is.vlabeled returns a vector/list element with logical values
}
\author{
Martin Elff
}

\seealso{
adj.measlev, is.vlabeled
}
\examples{
\dontrun{
## input of adj.measlev
is.labeled(myvector)
}
}
\keyword{internal}
