\name{pisa.country}
\alias{pisa.country}
\docType{data}
\title{PISA country abbreviations and ISO codes
}
\description{Contains PISA participating countries, names, abbreviations, and ISO codes 
}
\usage{data(pisa.country)}
\format{
  A data frame with 243 observations on the following 3 variables.
  \describe{
    \item{\code{Country}}{a character vector}
    \item{\code{ISO}}{a character vector}
    \item{\code{Code}}{a numeric vector}
  }
}

\source{PISA Technical Report
}

\examples{
data(pisa.country)
str(pisa.country) 
}
\keyword{datasets}
