\name{intsvy-package}
\alias{intsvy-package}
\alias{intsvy}
\docType{package}
\title{International Assessment Data Manager}
\description{Provides tools for importing, merging, and analysing data from international assessment studies (TIMSS, PIRLS, PISA, and PIAAC and others)}
\details{
\tabular{ll}{
Package: \tab intsvy\cr
Type: \tab Package\cr
Version: \tab 2.3\cr
Date: \tab 2018-10-09\cr
License: \tab GPL-2\cr
}
intsvy allows useRs to work with international assessment data (e.g., TIMSS, PIRLS, PISA, ICILS, and PIAAC). The data is publicly available @ rms.iea-dpc.org and www.oecd.org/pisa in SPSS format. Data and merge functions print variable labels and the name of participating countries in international assessments as well as import data directly into R for the variables in student, parent, school, and teacher instruments and countries selected by the useR. Analysis functions, including mean statistics, standard deviations, regression estimates, correlation coefficients, and frequency tables, calculate point estimates and standard errors that take into account the complex sample design (i.e., replicate weights) and rotated test forms (i.e., plausible achievement values). 
}

\author{Daniel Caro <dcarov@gmail.com>, Przemyslaw Biecek <przemyslaw.biecek@gmail.com>}
\references{PISA, PIAAC, PIRLS, and TIMSS Technical Reports}